refactor: architectural improvements to reduce complexity and improve modularity
- Extract frontend-only types from store.ts to dedicated types.ts module, eliminating upward dependencies from offline cache modules
- Deduplicate detail-fetch pattern: openThingDetail, navigateThingDetail, and goBackThingDetail now share loadAndSetDetail helper
- Remove dead exports: HealthResponseSchema from schemas.ts and getCachedThingsByType from cache-things.ts
- Optimize getPendingCount to use IDB count() instead of loading all records into memory
- Update cache-calendar.ts to import types from types.ts instead of store.ts
