+++ frontend/src/offline/cache-things.ts
@@ -1,5 +1,4 @@
 import type { Thing } from '../generated/api-types'
-import type { TypeHint } from '../utils'
 import { putAll, getAll, clearStore } from './idb'
 
 /**
@@ -18,10 +17,3 @@ export async function getCachedThings(): Promise<Thing[]> {
   return getAll('things')
 }
 
-/**
- * Retrieve cached things filtered by type_hint.
- */
-export async function getCachedThingsByType(typeHint: TypeHint): Promise<Thing[]> {
-  const all = await getAll('things')
-  return all.filter(t => t.type_hint === typeHint)
-}
