+++ frontend/src/store.ts
@@ -41,8 +41,6 @@ import { z } from 'zod'
 import type {
   Thing,
   ThingType,
-  CallUsage,
-  ModelUsage,
   Relationship,
   RequestyModel,
   ProactiveSurface,
@@ -52,7 +50,6 @@ import type {
   LearnedPreference,
   MorningBriefing,
   BriefingPreferences,
-  Nudge,
   WeeklyBriefing,
   ModelSettings,
   UserSettings,
@@ -61,131 +58,20 @@ import type {
   ConnectionSuggestion,
 } from './generated/api-types'
 
-export type { Thing, ThingType, ModelUsage, Nudge }
+// Re-export all shared types from types.ts so existing `import { useStore, type Foo } from './store'` still works
+export type {
+  Thing, ThingType, ModelUsage, Nudge, TypeHint,
+  ChatSession, WebSearchResult, ContextThing, GmailMessage, ReferencedThing,
+  AppliedChanges, BriefingItem, BriefingStats, CalendarEvent, CalendarStatus,
+  GmailStatus, InteractionStyle, ChatMode, StreamingStage, SessionStats,
+  ChatMessage, AuthUser,
+} from './types'
 
-// ── Frontend-only types (not derived from Pydantic models) ────────────────────
-
-export interface ChatSession {
-  id: string
-  title: string
-  origin: string | null
-  created_at: string
-  last_active_at: string
-  message_count: number
-}
-
-export type { TypeHint } from './utils'
-
-export interface WebSearchResult {
-  title: string
-  url: string
-  snippet: string
-}
-
-export interface ContextThing {
-  id: string
-  title: string
-  type_hint?: string | null
-}
-
-export interface GmailMessage {
-  id: string
-  subject: string
-  from: string
-  date: string
-  snippet: string
-}
-
-
-export interface ReferencedThing {
-  mention: string
-  thing_id: string
-}
-
-export interface AppliedChanges {
-  created?: { id: string; title: string; type_hint?: string }[]
-  updated?: { id: string; title: string; [key: string]: unknown }[]
-  deleted?: string[]
-  context_things?: ContextThing[]
-  referenced_things?: ReferencedThing[]
-  web_results?: WebSearchResult[]
-  gmail_context?: GmailMessage[]
-  calendar_events?: CalendarEvent[]
-}
-
-// BriefingItem.thing is dict[str, Any] in backend but always a Thing in practice
-export interface BriefingItem {
-  thing: Thing
-  importance: number
-  urgency: number
-  score: number
-  reasons: string[]
-}
-
-export interface BriefingStats {
-  active_things: number
-  checkin_due: number
-  overdue: number
-}
-
-export interface CalendarEvent {
-  id: string
-  summary: string
-  start: string
-  end: string
-  all_day: boolean
-  location: string | null
-  status: string
-}
-
-export interface CalendarStatus {
-  configured: boolean
-  connected: boolean
-}
-
-export interface GmailStatus {
-  connected: boolean
-  email: string | null
-}
-
-export type InteractionStyle = 'auto' | 'coach' | 'consultant'
-
-export type ChatMode = 'normal' | 'planning'
-
-export type StreamingStage = 'context' | 'reasoning' | 'response' | null
-
-export interface SessionStats {
-  prompt_tokens: number
-  completion_tokens: number
-  total_tokens: number
-  api_calls: number
-  cost_usd: number
-  per_model: ModelUsage[]
-}
-
-export interface ChatMessage {
-  id: number | string
-  session_id: string
-  role: 'user' | 'assistant' | 'system'
-  content: string
-  applied_changes: AppliedChanges | null
-  questions_for_user: string[]
-  prompt_tokens?: number
-  completion_tokens?: number
-  cost_usd?: number
-  model?: string | null
-  per_call_usage?: CallUsage[]
-  timestamp: string
-  streaming?: boolean
-  streamingStage?: StreamingStage
-}
-
-export interface AuthUser {
-  id: string
-  email: string
-  name: string
-  picture: string | null
-}
+import type {
+  AuthUser, BriefingItem, BriefingStats, CalendarEvent, CalendarStatus,
+  ChatMessage, ChatMode, ChatSession, GmailStatus, InteractionStyle,
+  Nudge, SessionStats,
+} from './types'
 
 interface ReliState {
   currentUser: AuthUser | null
@@ -430,6 +316,20 @@ async function fetchThingDetailWithFallback(
   }
 }
 
+function loadAndSetDetail(
+  id: string,
+  get: () => ReliState,
+  set: (partial: Partial<ReliState>) => void,
+) {
+  fetchThingDetailWithFallback(id).then(([thing, rels]) => {
+    if (get().detailThingId === id) {
+      set({ detailThing: thing, detailRelationships: rels, detailLoading: false })
+    }
+  }).catch(() => {
+    if (get().detailThingId === id) set({ detailLoading: false })
+  })
+}
+
 export const useStore = create<ReliState>((set, get) => ({
   currentUser: null,
   authChecked: false,
@@ -630,13 +530,7 @@ export const useStore = create<ReliState>((set, get) => ({
 
   openThingDetail: (id: string) => {
     set({ detailThingId: id, detailHistory: [], detailLoading: true, detailThing: null, detailRelationships: [] })
-    fetchThingDetailWithFallback(id).then(([thing, rels]) => {
-      if (get().detailThingId === id) {
-        set({ detailThing: thing, detailRelationships: rels, detailLoading: false })
-      }
-    }).catch(() => {
-      if (get().detailThingId === id) set({ detailLoading: false })
-    })
+    loadAndSetDetail(id, get, set)
   },
 
   navigateThingDetail: (id: string) => {
@@ -649,13 +543,7 @@ export const useStore = create<ReliState>((set, get) => ({
       detailThing: null,
       detailRelationships: [],
     }))
-    fetchThingDetailWithFallback(id).then(([thing, rels]) => {
-      if (get().detailThingId === id) {
-        set({ detailThing: thing, detailRelationships: rels, detailLoading: false })
-      }
-    }).catch(() => {
-      if (get().detailThingId === id) set({ detailLoading: false })
-    })
+    loadAndSetDetail(id, get, set)
   },
 
   goBackThingDetail: () => {
@@ -663,13 +551,7 @@ export const useStore = create<ReliState>((set, get) => ({
     if (history.length === 0) return
     const prevId = history[history.length - 1]!
     set({ detailThingId: prevId, detailHistory: history.slice(0, -1), detailLoading: true, detailThing: null, detailRelationships: [] })
-    fetchThingDetailWithFallback(prevId).then(([thing, rels]) => {
-      if (get().detailThingId === prevId) {
-        set({ detailThing: thing, detailRelationships: rels, detailLoading: false })
-      }
-    }).catch(() => {
-      if (get().detailThingId === prevId) set({ detailLoading: false })
-    })
+    loadAndSetDetail(prevId, get, set)
   },
 
   closeThingDetail: () => {
